requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
dnspython==2.6.1
email_validator==2.1.2
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.14.0
Flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.5
packaging==24.0
pefile==2023.2.7
platformdirs==4.2.2
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
python-dotenv==1.0.1
python-multipart==0.0.9
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
setuptools==70.0.0
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.26.2
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.2
